Created the booking form
Created a material-ui themed form, along with validation.
When opening, the submit button is valid, but validation for empty fields occurs on submitting.
If errors are present, the form cannot be submitted.
Validation also occurs as the form is filled in, and checks for things like valid email address, and location/length combinations work.

Submitting form correctly writes the data to firebase.

NOT YET IMPLEMENTED:
Handling/displaying errors on the firebase functions API call.