Inside each expansion panel, place a disabled booking form.
Achieved this by passing the booking object to the BookingForm component. Inside the BookingFormComponent, check for this prop, and if it exists map the object to the object used to set the form values. Otherwise treat it as it always did.
Introduced state 'editing' which is initialised as true if the booking object was provided.
Also changed the date to use the form values - YET TO COMPLETELY CLEAN THIS UP.