Created Roles
All users are added into firestore on creation, with default roles set to empty object.
When getting user inside withAuthorization and withAuthentication HOC's, merge the firestore user and the firebase user.
AuthUserContext can provide the user at any point in the app, and components can be hidden/shown based on the roles values.