Environment configuration
Finally setup the env stuff!
When deploying, created a predeploy script which checks which project is in use, and then runs either the 'prod' or 'dev' script in the package.json.
These scripts inject a REACT_APP_ENV environment variable, then runs a build, which uses the env variables to determine which firebase config to use.
The additional benefit of this approach is that you cannot accidentally deploy the dev environment into production or vice versa.