Integrated acuity and stripe using firebase functions
Created firebase function 'sidebar', which serves html to acuity when viewing an appointment.
This function searches for the form field 'invoice id'. If the field doesn't, appointment doesn't support invoices.
If it does... if the value is empty, invoice is not sent, and provide option to send invoice. If it isn't empty, lookup the invoice status and display it.

When hitting 'send invoice' it calls another firebase function 'sendInvoice'.
This will search stripe for a customer with the given email, or create a new customer if not found.
It then creates an invoice item for the science club, creates the invoice, then sends it off.