Added typescript to project
Started by uplifting ConfirmationDialog HOC. Since its a higher-order-component it was a nice starting point to test it out.
Everything looks good - compile time is slightly slower on change... would not want that to slow as i uplift more components.

Issue with HOC is that the props need to be explicitly imported by the class being wrapped, and '&' it along with that components own props. It doesn't come as part of the props by default.