Introduced TypeScript into firebase functions
Refactored the acuity and stripe integration into typescript. Pulled sidebar and sendInvoice functions into own file to achieve this.
Using @types/stripe, however no types exist for acuity, so made my own inside types/acuity/index.d.ts, which only includes the minimal types that are used by the project.

Added HTML response for case that invoice is created, but fails to save back into acuity. If so, do not send the invoice, but tell the user the invoice has been created, and to contact Ryan.