When customer submits form, update firestore and trigger emails
// IN STORE ONLY //
Added the on form submit trigger to provide the values to the new firebase function onFormSubmit.
Function searches for the booking, and updates firestore with values, before asking apps script to send confirmation and notification emails.
If booking is not found, go to apps script to send error email.

Mobile form yet to be integrated.