Added separate fun facts field
This was added because by appending it to the questions, it was causing issues with the notification emails.
To fix the notification email issues would mean being able to split out the fun facts, so making it its own field made most sense.
Also updated functions endpoint to australia-southeast1 in the form submission apps script.