Passed environment into apps script
This is a workaround for not having a separate development and production environment for apps-script... which isn't quite possible. Ideally apps-script should be ditched, but its heavily used for now.
This patch tells apps script if its dev or prod, and that can be used to determine which calendar to use, test calendars or prod calendars for managing bookings in Google Calendar.