Moved booking constants inside 'functions' dir
And turned this directory into an npm package called 'fizz-kidz'. This way, the package can be installed as a local package in React, and just imported normally from within functions. This allows for a single source of truth for booking form values.
Added new creation values to match the new form, which matches the new website packages. Added 6th question of Creations.