Added price selection when sending an invoice
Added a `listItems` prop into the `ConfirmationDialog` which, when provided, will show a list of the items along with the title. If this prop is provided, the `onConfirm()` callback will also provide the value of the selected list item.
This extended confirmation dialog was used to show different prices available for selection when sending an invoice.
Also fixed validation of params when sending an invoice - this was broken and previously no appointmentTypeID was being provided, and all classes of the client were being updated, including previous terms.