Separated booking types into Booking, FormBooking and FirestoreBooking
This is because when retrieving from Firestore, the dateTime value is a firestore.Timestamp. However when writing to firestore it can just be date object. So passing around a firestore.Timestamp object is not needed.
Furthermore, passing around a firestore.Timestamp object is a headache, since when parsed into JSON it becomes an object of `{ seconds: number: nanoseconds: number }` which is gross.

However a conversion is still needed inside the `updateBooking` firebase function, since a `Date` object is still passed as a string when serializing, and then needs to be serialized into a `Date` so that when written to firestore it will be stored as a `Timestemp`.... hmmm..... it all seems frustratingly convoluted.
Perhaps once I uplift cloud functions that manages bookings into typescript I might come up with a nice way to handle all this.

This commit also fixes the accidental removal of `eventId`, and reminded me why a merge was done in the first place 🤦‍♂️