Added check to ensure form confirmation emails only sent once
Due to issues where iPhones can resubmit forms when re-opening safari, a form can be resubmitted hours later, and a parent gets a confirmation email again.
This patch adds a check that the creation1 value has not been set before, and if so, sends the emails. Otherwise, it will just update the booking but no need to notify anyone of anything.