Added ability to void an invoice, and resend a new one to backend only
Pulled out the sending of invoice from within the `sendInvoice()` firebase function into its own method, and reused it within a new method `voidAndResendInvoice()`.
This function will essentially do the exact same thing, however first retrieve the invoiceId from the appointment, void it, and then issue a new one.
When invoice is sent, the calll to `retrieveInvoiceStatus()` needs to be made again, however the new method is called from the menu which is separated entirely from the `RetrieveInvoiceStatusWithAction` component. To overcome this, I created a `useForceRerenderComponent` hook, which is used in the `ExpandableTableRow` to rerender the entire row when a new invoice is sent. This will re-trigger the call to retrieve the invoice status, and get the latest non-void invoice.

Currently only the backend is implemented, with the front using hard coded values.
Next step will be to select the price from a confirmation dialog like used when sending and invoice.