Created mjml firebase function
This is because the official mjml API is very unreliable. Instead, created a firebase function that uses the mjml npm package to convert stuff ourselves.
However doing this uncovered some issues in the mjml templates, which had to be fixed.

Updated apps-script to call the new firebase function. Included environment to choose whether to use prod or dev firebase function.