Store each individual holiday program booking in firestore under payment intent id
This is needed to ensure the synchronous nature of paying and booking into acuity.
As soon as a payment is attempted, store all information in firestore.
Upcoming work will create a webhook to stripe to listen for payment intent events, and then handle it.
Ie if payment succeeds, then book into acuity. If it fails, send an emailing to the customer telling them.