created pubsub for onboarding form submission
this is because the function does too much, and cannot be completed with the 5 second timeout that paperforms expects a response. previously i returned a 200 response immediately to avoid this, however it seems firebase functions reads the response codes and begins cleanup before my function is finished.

to get around this, i moved heavy lifting into a pubsub which i trigger from within the form submission.

also refactored employee types to properly reflect their possible states, and remove not null assertions from everywhere