actually used react-router
restructured the routes object to properly reflect how the application works.
created a root layout which is where all providers wrap the app.
created a dashboard layout to share the nav between all pages within the dashboard.
removed routes constants - this is because using them to navigate to and from is kinda of useless, since react router uses relative routes... so no clean way to do this.
instead of using the react-router lazy method, use React.lazy for lazy loading routes. this allows using <Suspense> which shows a loader while waiting for the component to load, instead of just freezing for a second.