separate users into either a customer or non customer
only staff can be members of an organisation, so when signing in to firebase, add a custom claim to the auth user that tells us if they are a customer. firestore security rules can protect the database based on this information.

created a 404 error boundary.
only show bookings of the current org on the bookings screen - all other screens are still full access across orgs.
immproved dashboard layout flex stuff.
removed custom roles, and instead use clerk roles and permissions to protect routes, and hide/show UI elements.
rebuilt sign in page.