collect child birthday in holiday program enrolment form.
this is instead of simply collecting their age. by doing this, we can now store their birthday into zoho, and create marketing flows for when that childs birthday approaches.
to make this work with zoho cleanly, the best approach was to simply create 5 children slot fields in zoho, and add the child into a free slot, if one doesn't exist with a matching birthday