joined trpc and all webhooks to a single 'api' cloud function
achieved this using express, and serve all trpc routes on a single '/api/trpc' endpoint.
this greatly reduces the number of cloud functions, so that there is only a single http cloud function left, the api.
the remaining functions are all pubsub functions.

i also consolidated the onboarding paperform webhook to be handled by the single paperform webhook,
and then let the actual work be done in the pubsub. this menat putting all the logic into a single place,
rather than doing some work in the webhook and the rest in its own dedicated pubsub - which was unecessary.